requirements.txt
together
python-dotenv